+++ modules/src/mock/mod.rs
@@ -1,5 +1,0 @@
-mod access_control;
-mod pauseable;
-
-pub use access_control::{MockModerated, MockModeratedDeployer, MockModeratedRef};
-pub use pauseable::PauseableCounter;